Теперь блоки не знают весь список блоков :)
