+++ app/src/main/java/com/example/hits_android/CodeBlocksScreen.kt
@@ -1,0 +1,25 @@
+package com.example.hits_android
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import com.example.hits_android.model.ReorderListViewModel
+
+@Composable
+fun CodeBlocksScreen(
+    vm: ReorderListViewModel
+) {
+    Box(
+        Modifier
+            .fillMaxHeight(0.93f)
+            .fillMaxWidth()
+    ) {
+        Column () {
+            FunctionSelection(vm = vm)
+            VerticalReorderList(vm = vm)
+        }
+    }
+}
