+++ app/src/main/java/com/example/hits_android/blocks/BreakBlock.kt
@@ -68,12 +68,11 @@ class BreakBlock(
     }
 
     @Composable
-    override fun BlockComposable(item: Block, codeBlocksList: List<Block>) {
+    override fun BlockComposable(item: Block) {
         Box(
             modifier = Modifier
                 .fillMaxSize()
                 .height(50.dp)
-                .padding(start = calculatePadding(codeBlocksList, item.key))
                 .clip(RoundedCornerShape(24.dp))
                 .background(MaterialTheme.colorScheme.errorContainer),
             contentAlignment = Alignment.Center
