+++ app/src/main/java/com/example/hits_android/blocks/FinishProgramBlock.kt
@@ -31,7 +31,7 @@ class FinishProgramBlock(
     }
 
     @Composable
-    override fun BlockComposable(item: Block, codeBlocksList: List<Block>) {
+    override fun BlockComposable(item: Block) {
         Box(
             modifier = Modifier
                 .fillMaxWidth(),
