+++ app/src/main/java/com/example/hits_android/blocks/FunctionBlock.kt
@@ -79,7 +79,7 @@ class FunctionBlock(
     }
 
     @Composable
-    override fun BlockComposable(item: Block, codeBlocksList: List<Block>) {
+    override fun BlockComposable(item: Block) {
 
     }
