+++ app/src/main/java/com/example/hits_android/blocks/IfBlock.kt
@@ -100,13 +100,12 @@ class IfBlock(
     }
 
     @Composable
-    override fun BlockComposable(item: Block, codeBlocksList: List<Block>) {
+    override fun BlockComposable(item: Block) {
         item as IfBlock
         Box(
             modifier = Modifier
                 .fillMaxSize()
                 .height(70.dp)
-                .padding(start = calculatePadding(codeBlocksList, item.key))
                 .clip(RoundedCornerShape(24.dp))
                 .background(MaterialTheme.colorScheme.surfaceVariant)
         ) {
