+++ app/src/main/java/com/example/hits_android/blocks/MainBlock.kt
@@ -32,7 +32,7 @@ class MainBlock(
     }
 
     @Composable
-    override fun BlockComposable(item: Block, codeBlocksList: List<Block>) {
+    override fun BlockComposable(item: Block) {
         Box(
             modifier = Modifier
                 .fillMaxWidth(),
