+++ app/src/main/java/com/example/hits_android/blocks/OutputBlock.kt
@@ -96,13 +96,12 @@ class OutputBlock(
     }
 
     @Composable
-    override fun BlockComposable(item: Block, codeBlocksList: List<Block>) {
+    override fun BlockComposable(item: Block) {
         item as OutputBlock
         Box(
             modifier = Modifier
                 .fillMaxSize()
                 .height(70.dp)
-                .padding(start = calculatePadding(codeBlocksList, item.key))
                 .clip(RoundedCornerShape(24.dp))
                 .background(MaterialTheme.colorScheme.surfaceVariant)
         ) {
